Make userpreference using player data class and insert user data to database

+++ app/src/main/java/com/catnip/rockpaperscissorchapter6and7/ui/game/mode/dialog/PlayerMenusPresenter.kt
@@ -37,7 +37,7 @@ class PlayerMenusPresenter(
                     Log.d("insertPlayer", "insertPlayer: $playerId")
                     view.onPlayerIDCallback(
                         Player(
-                        id = playerId.toInt(),
+                        id = playerId,
                         name = player.name
                     ))
                 }
